Rearranged wind data to wind vector and adding all hrrr met data to process data. Will likely change names/structure a bit to make things more clear before opening PR back to Veloserver.
