+++ Dockerfile
@@ -31,6 +31,7 @@ RUN apt-get update && apt-get install -y \
     amqp-tools \
     openssh-client \
     gfortran \
+    gdal-bin \
     --no-install-recommends && rm -r /var/lib/apt/lists/* \
     && wget https://ftp.cpc.ncep.noaa.gov/wd51we/wgrib2/wgrib2.tgz \
     && mkdir -p /usr/local/grib2/ \
